omni-master/ai-engine/requirements.txt
flask==3.0.3
flask-cors==4.0.1
imagehash==4.3.1
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
